autocompiler now serves index files instead of 404s
When requesting a path for which no page/asset rep exists, the autocompiler will try to serve the path if a file exists there and the path does not end with a slash. If the path does end with a slash, it'll try to serve one of the index files instead.

--HG--
branch : 2.1