added #filter! and #layout! methods to ItemRep
These two methods perform basic filtering and layouting. The old #do_filter and #do_layout methods still exist (and now use the new methods), but will be removed. The new #filter! and #layout! methods are also the ones that will be called by the (not yet written) DSL.

--HG--
branch : experimental