autocompiler now serves asset content correctly
Due to the request path being cleaned (String#cleaned_path), the autocompiler would never be able to find the rep for the requested asset (because it'd end with a slash, and asset paths rarely end with a slash). Instead, the autocompiler would serve the already compiled content from a previous "nanoc co" call. Moral of the story: never use #cleaned_path on disk paths or web paths.