Add support for `auto_remove_stale_items` config
This allows the `/output` directory to act as a build directory, as all stale files are removed after compilation (see issue #1) …

It's prob'ly not the most efficient way to accomplish this, and I'm positive the way I invoked the `clean_strays` command is wrong, but it does the trick. @ddfreyne, what's the right way to do this?