Avoid libxml2 encoding bug.
Calling inner_html= resulted in all kinds of encoding weirdness due to a bug in
a common version of libxml2. Building a tree of nodes and using children=
instead works properly.

The bug is present in the version of libxml2 that ships with OS X Lion, so
working around this in Nanoc itself should save a lot of users a good bit of
trouble.

References:
https://github.com/tenderlove/nokogiri/issues/458