Support filtering Items Arrays by Regex
This is a convenience API for working with the `@items` array in nanoc's
`preprocess` block. In that context, it is common to manipulate items
according to a shared part of their identifier.

For example, with this change in place, users could succinctly create a
collection of all authors for a given site's weblog:

    @items[%r{^/weblog/authors/}].each { |author|
      # etc...
    }