Remove Inch badge
nanoc is a tool, not a library. As much as I care about documentation,
API documentation is not where it’s at for nanoc. Because of this, I
don’t feel that the Inch badge belongs in the nanoc README.

I will continue to use Inch for other projects where it is more
appropriate (for instance, ddfreyne/cri).

CC @rrrene