Add DateTime handling to attribute_to_time
This enhances `attribute_to_time` which previously removed all time
information from a DateTime. This happened because DateTime identifies
itself as a Date.

    2.2.3 :001 > require 'date'
     => true
    2.2.3 :002 > DateTime.now.is_a?(Date)
     => true

Up to now, if `time.is_a?(Date)` is true, `attribute_to_time` will
create a new Time, with just year, month and day set. To fix that
behaviour, this commit adds a new condition at the beginning of
`attribute_to_time`, which uses `time.to_time` as long as `time` is a
DateTime.