Optimize the handling of binary items
When writing large binary items (e.g. audio or video files), creating a
hard link pointing to the original avoids duplicating the data on disk.
This is specially useful for large binary items.

Hard linking will fail when the source file is not on the same device as
the output directory (e.g. when /tmp is a tmpfs). In that case, fallback
to the previous behavior of copying, what does duplicate the space, but
should always work.