Prevent race condition in diff generation
The call to `Nanoc::Core::NotificationCenter.sync` helped avoid some issues, but it is still possible for the diff generation to fail because the original file is no longer available. This happens because `.sync` unblocks as soon as the last notification is picked up, rather than when the last notification is finished.