Fix nanoc-cli build
It turns out that nanoc-cli was not being tested at all, which explains why I was so successful in setting it up.

Unfortunately, fixing it was not trivial, and so this single commit contains a handful of unrelated changes:

* Turn on nanoc-cli tests

* Merge Nanoc::Spec into common/. nanoc-cli needs to use it, and it’d require a dependency onto nanoc, which is not what we want.

* ShowPlugins (part of nanoc-cli) no longer knows about Deployer, but there is now support for adding custom plugin classes to ShowPlugins.

* ShowRules is moved from nanoc-cli back to nanoc, because it requires knowledge of the rules DSL, which is not part of nanoc-cli or nanoc-core.

This isn’t the way I intend to make changes, and I’d rather have rolled back, but in this case I felt that rolling back would have been difficult and not have gained us as much as rolling forward. Consider this commit very much an exception to the rule, though.