Read files as UTF-8 for diff generation
Nanoc always writes output files in UTF-8, so when reading them, they need to be interpreted as UTF-8 as well. Otherwise, depending on the default external encoding, the diff generation might fail later on (confusingly, on Windows, it just locks up the Ruby process -- no exception).

This change lacks a test case (for now), as I don’t see a way of testing this now.