Fix uninitialized constant/undefined method bug with guard-nanoc 2.1.7
The primary fix is making the constants absolute, but additionally, I’ve added a test to verify the behavior of the `live` command for real, which should prevent this from happening again.

The __NANOC_DEV_LIVE_DISABLE_VIEW is a hack to disable `nanoc view` during the tests, so that the separate thread doesn’t raisen an exception and clog up the test output.