nanoc-cli: Only try to clean UTF-8 chars from UTF-8 strings
Previously, when attempting to load a custom filter in a Windows Nanoc
installation, Nanoc was crashing with indications that the UTF-8 stream cleaner
was unable to process the supplied string.
Now, we only attempt to clean UTF-8 characters out of the string if it is UTF-8
encoded.

Should fix #1502. (Including the contracts gem issue).