Make Checksummer faster and more correct
The new implementation is faster because of two reasons:

- It more aggressively avoids calculating checksums of already-calculated items
- It effectfully updates a hash rather than using Immutable::Set

It is also more correct because it replaces use of <recur> with a numerical reference to a previously-seen object.

Each newly checksummed object will have a number suffix, e.g.
`Array#5<…>`. Each previously seen object will have an @-reference, e.g.
`Array#9<@4,@5,>`.

This will invalidate existing stored checksums in Nanoc sites, but it is unavoidable and worth the tradeoff.