Pull in non-enqueued reps if they are useful
When an item rep fails to be compiled due to a dependency, enqueue all
the dependent reps as well. This is a heuristic that could drastically
improve performance.

Even though the dependent reps might not have changed, recompiling them
is quick because the compiled content cache will be used.

This change should drastically improve the compilation performance of
sitemaps, which tend to depend on every item in a site.